test: divide components; add test for new components;
