+++ src/constants/index.ts
@@ -1,2 +1,10 @@
 export const MAX_ITEMS_PER_PAGE = 3;
 export const SLICE_LENGTH = 2;
+
+export const DURATION_TIME_KEY = '--duration-time';
+
+export const ArrowUpKey = 'ArrowUp';
+export const ArrowDownKey = 'ArrowDown';
+
+export const MIN_RATE = 0.25;
+export const MAX_RATE = 5;
